pyproject.toml
[tool.poetry]
name = "lawyer"
version = "0.1.0"
description = ""
authors = ["Mapc <mapctrigotixd@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
python-dotenv = "^0.19.2"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.0.0"
djangorestframework = "^3.13.1"
pyTelegramBotAPI = "^4.4.0"
celery = "^5.2.3"
redis = "^4.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
